+++ mukurobot/checkip.py
@@ -1,23 +1,24 @@
 from types import FunctionType
-import requests
+import aiohttp
 import logging
 
 _log : logging.Logger = logging.getLogger(__name__)
 _log.setLevel(logging.INFO)
 
-def check_ip(blacklist_func=(lambda x:False), /, *, endpoint='https://ipinfo.io/json') -> None:
-    resp = requests.get(endpoint)
+async def check_ip(blacklist_func=(lambda x:False), /, *, endpoint='https://ipinfo.io/json') -> None:
+    async with aiohttp.ClientSession() as client:
+        resp = await client.get(endpoint)
 
-    if resp.status_code != 200:
-        raise RuntimeError(f'API returned status {resp.status_code}')
+        if resp.status != 200:
+            raise RuntimeError(f'API returned status {resp.status_code}')
 
-    ip = resp.json()['ip']
+        ip = (await resp.json())['ip']
 
-    _log.info(f'IP check requested (your IP: {ip})')
+        _log.info(f'IP check requested (your IP: {ip})')
 
-    if blacklist_func(ip):
-        _log.error(f'IP address disallowed to start connections')
-        raise ValueError(f'IP address disallowed to start connections')
+        if blacklist_func(ip):
+            _log.error(f'IP address disallowed to start connections')
+            raise ValueError(f'IP address disallowed to start connections')
 
 def blacklist_from_txt_file(file):
     def func(ip) -> bool: